more cleanup

+++ aphorismServer/product/talk.c
@@ -58,11 +58,11 @@ int main(int argc, char *argv[])
     // send the user's input lines to the server
     bzero(buffer, MAXBUF);
     while (1) {
-      int n = MAXBUF;
+      size_t n = MAXBUF;
       int got = getline (&buf, &n, stdin);
       if (got < 0) break;
       buf[got] = 0;
-       printf ("send to %s:%d %s\n",  inet_ntoa(dest.sin_addr), ntohs(dest.sin_port), buf);
+      printf ("send to %s:%d %s\n",  inet_ntoa(dest.sin_addr), ntohs(dest.sin_port), buf);
       sendto (sockfd, buf, got+1, 0, (struct sockaddr *)&dest, sizeof(dest));
       printf ("sent %d bytes\n", got);
     }
